+++ GameParticles/ParticleEmitter.h
@@ -12,27 +12,27 @@ class ParticleEmitter
 public:
 	ParticleEmitter();
 	~ParticleEmitter();
-	
+
 	void SpawnParticle();
 	void update();
 	void draw();
 
-	void addParticleToList(Particle *p );
-	void removeParticleFromList( Particle *p );
+	void addParticleToList(Particle *p);
+	void removeParticleFromList(Particle *p);
 
 	void Execute(Vect4D& pos, Vect4D& vel, Vect4D& sc);
 
 private:
-	
+
 	Vect4D	start_position;
 	Vect4D	start_velocity;
-	
-	float	max_life;
+
+	double	max_life;
 	int		max_particles;
 	int		last_active_particle;
-	float	spawn_frequency;
-	float	last_spawn;
-	float	last_loop;
+	double	spawn_frequency;
+	double	last_spawn;
+	double	last_loop;
 
 
 	// added for speed efficiency
@@ -40,7 +40,7 @@ private:
 	Particle *headParticle;
 	Vect4D	vel_variance;
 	Vect4D	pos_variance;
-	float	scale_variance;
+	double	scale_variance;
 	std::list<Particle> drawBuffer;
 	std::list<Particle> particle_list;
